smartfridge-dataset/requirements.txt
opencv-python
Pillow
numpy
google-cloud-vision
pytesseract
pandas
dateparser
python-dotenv
paddlepaddle==2.6.3
paddleocr
# Note: paddlepaddle install differs depending on CPU/GPU; see official docs
